Introduce `BuildSrcProjectConfigurationAction`
This new extension point allows any module in the classpath to
contribute configuration actions targeting specifically at the `buildSrc`
root project.

The motivating use-case is allowing gradle-script-kotlin to contribute
tasks and source-sets to the `buildSrc` project without introducing
a hard-coded dependency (see gradle/gradle-script-kotlin#230).

As a bonus, the Groovy buildSrc configuration was extracted to its own
configuration action class and moved closer to `GroovyPlugin` so it
could be applied by type instead of id.